pyproject.toml
[project]
name = "agrinote-scraper-for-lambda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "awslambdaric>=4.0.0",
    "boto3>=1.42.54",
    "gspread>=6.2.1",
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-playwright>=0.7.2",
]
